requirements.txt
neo4j>=5.20.0
python-dotenv>=1.0.1
pandas>=2.1.4
numpy>=1.26.2
sentence-transformers>=2.2.2
scikit-learn>=1.3.2
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.21
langchain-huggingface>=0.0.8
langchain-ollama>=0.1.0
huggingface_hub>=0.23.0
streamlit>=1.28.0
plotly>=5.18.0
networkx>=3.2.1
tqdm>=4.66.1
pytest>=7.4.4
matplotlib>=3.8.0
